requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
requests==2.32.0
PyYAML==6.0
mysql-connector-python==8.1.0